FEC-11250 Exoplayer upgrade v2.14.0 (#713)

+++ playkit/src/main/java/com/kaltura/android/exoplayer2/dashmanifestparser/CustomSingleSegmentIndex.java
@@ -50,12 +50,12 @@ import com.kaltura.android.exoplayer2.source.dash.manifest.RangedUri;
     }
 
     @Override
-    public int getSegmentCount(long periodDurationUs) {
+    public long getSegmentCount(long periodDurationUs) {
         return 1;
     }
 
     @Override
-    public int getAvailableSegmentCount(long periodDurationUs, long nowUnixTimeUs) {
+    public long getAvailableSegmentCount(long periodDurationUs, long nowUnixTimeUs) {
         return 1;
     }
